Refactor: TokenManager module 변경 (domain)
