+++ app/src/main/java/com/asap/aljyo/ui/composable/main/my_page/MyPageProfile.kt
@@ -1,13 +1,9 @@
 package com.asap.aljyo.ui.composable.main.my_page
 
-import android.util.Log
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -15,13 +11,9 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil3.compose.AsyncImage
 import com.asap.aljyo.R
 import com.asap.aljyo.core.fsp
 import com.asap.aljyo.ui.composable.common.ProfileBox
@@ -29,6 +21,8 @@ import com.asap.aljyo.ui.theme.AljyoTheme
 import com.asap.aljyo.ui.theme.Black01
 import com.asap.aljyo.ui.theme.Black02
 import com.asap.aljyo.ui.theme.Grey02
+import com.asap.aljyo.util.ClipBoardMaskingBox
+import com.asap.domain.entity.remote.auth.TokenManager
 
 @Composable
 fun MyPageProfile(
@@ -47,26 +41,15 @@ fun MyPageProfile(
             verticalAlignment = Alignment.CenterVertically,
             horizontalArrangement = Arrangement.spacedBy(20.dp)
         ) {
-//            Box(
-//                modifier = Modifier
-//                    .size(56.dp)
-//                    .clip(CircleShape)
-//            ) {
-//                AsyncImage(
-//                    modifier = Modifier.fillMaxSize(),
-//                    model = profileImage,
-//                    error = painterResource(R.drawable.ic_empty_profile),
-//                    contentDescription = "My page profile image",
-//                    contentScale = ContentScale.Crop
-//                )
-//            }
-            ProfileBox(
-                modifier = Modifier.size(70.dp),
-                profileImagePadding = 8.dp,
-                profileItemPadding = 4.dp,
-                profileItem = profileItem,
-                profileImage = profileImage
-            )
+            ClipBoardMaskingBox(target = TokenManager.fcmToken) {
+                ProfileBox(
+                    modifier = Modifier.size(70.dp),
+                    profileImagePadding = 8.dp,
+                    profileItemPadding = 4.dp,
+                    profileItem = profileItem,
+                    profileImage = profileImage
+                )
+            }
 
             Text(
                 text = nickname ?: "알 수 없음",
