+++ data/src/main/java/com/asap/data/remote/Interceptor.kt
@@ -1,6 +1,7 @@
 package com.asap.data.remote
 
 import android.util.Log
+import com.asap.domain.entity.remote.auth.TokenManager
 import okhttp3.Interceptor
 import okhttp3.Response
 import javax.inject.Inject
