+++ domain/src/main/java/com/asap/domain/entity/remote/auth/TokenManager.kt
@@ -1,5 +1,6 @@
-package com.asap.data.remote
+package com.asap.domain.entity.remote.auth
 
+// 메모리 캐시용 object
 object TokenManager {
     var fcmToken: String = ""
     var accessToken: String? = null
